+++ src/main/java/com/example/demo/security/TokenProvider.java
@@ -1,7 +1,7 @@
 package com.example.demo.security;
 
 
-public interface AuthService {
+public interface TokenProvider {
     String generateToken(String username);
     boolean validateToken(String token);
     String extractUsernameFromToken(String token);
