+++ src/main/resources/application.properties
@@ -33,5 +33,6 @@ spring.flyway.schemas=url_shortener
 logging.level.org.flywaydb.core=DEBUG
 logging.level.org.springframework.security=DEBUG
 
-#
-spring.cache.cache-names=authCache
+
+# todo/ dev-only: Must be changes and added to environment variables
+jwt.secret=SECRET_KEY_FOR_JWT
