+++ src/test/resources/application-test.properties
@@ -12,3 +12,5 @@ spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.use_sql_comments=true
 spring.jpa.properties.hibernate.type=trace
 spring.jpa.properties.hibernate.highlight_sql=true
+
+
